se creó función para editar data en kitchen
