+++ src/data/data.json
@@ -74,6 +74,91 @@
     }
   ],
   "orders": [
-    
+    {
+      "waiter": "luis_mesero@burgerqueen.com",
+      "clientName": "Jorge ",
+      "status": "pending",
+      "orderProducts": [
+        {
+          "id": 3,
+          "name": "Café con leche",
+          "price": 10,
+          "type": "desayuno",
+          "quantity": 1
+        },
+        {
+          "id": 4,
+          "name": "Sandwich de jamón y queso",
+          "price": 7,
+          "type": "desayuno",
+          "quantity": 1
+        },
+        {
+          "id": 2,
+          "name": "Jugo de Naranja",
+          "price": 7,
+          "type": "desayuno",
+          "quantity": 1
+        }
+      ],
+      "date": "30/11/2021 10:20:28",
+      "total": 24,
+      "id": 1
+    },
+    {
+      "waiter": "luis_mesero@burgerqueen.com",
+      "clientName": "Camila",
+      "status": "listo",
+      "orderProducts": [
+        {
+          "id": 6,
+          "name": "Hamburguesa doble",
+          "price": 15,
+          "type": "comida",
+          "quantity": 1
+        },
+        {
+          "id": 9,
+          "name": "Agua 500ml",
+          "price": 5,
+          "type": "comida",
+          "quantity": 1
+        },
+        {
+          "id": 7,
+          "name": "Papas fritas",
+          "price": 5,
+          "type": "comida",
+          "quantity": 1
+        }
+      ],
+      "date": "30/11/2021 10:20:51",
+      "total": 25,
+      "id": 2
+    },
+    {
+      "waiter": "armando_cocina@burgerqueen.com",
+      "clientName": "sofia",
+      "status": "pending",
+      "orderProducts": [
+        {
+          "id": 1,
+          "name": "Café americano",
+          "price": 5,
+          "type": "desayuno",
+          "quantity": 1
+        },
+        {
+          "id": 4,
+          "name": "Sandwich de jamón y queso",
+          "price": 7,
+          "type": "desayuno",
+          "quantity": 1
+        }
+      ],
+      "date": "30/11/2021 12:37:44",
+      "total": 12,
+      "id": 3
+    }
   ]
 }
