+++ src/helpers/helpHttp.js
@@ -36,10 +36,15 @@ export const helpHttp = () => {
         options.method = "DELETE";
         return customFetch(url, options);
     };
+    const patch = (url, options = {}) => {
+        options.method = "PATCH";
+        return customFetch(url, options);
+    };
     return {
         get,
         post,
         put,
         del,
+        patch,
     };
 };
